# backend/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "SafeBill RAG API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/reminders": {
      "get": {
        "tags": [
          "safebill-rag"
        ],
        "summary": "List Reminders",
        "operationId": "list_reminders_api_v1_reminders_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          },
          {
            "name": "days_ahead",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 60,
              "title": "Days Ahead"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 200,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemindersResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications": {
      "get": {
        "tags": [
          "safebill-rag"
        ],
        "summary": "List Notifications",
        "operationId": "list_notifications_api_v1_notifications_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          },
          {
            "name": "include_read",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Read"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/preferences": {
      "get": {
        "tags": [
          "safebill-rag"
        ],
        "summary": "Get Notification Preferences",
        "operationId": "get_notification_preferences_api_v1_notifications_preferences_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPreferenceView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "safebill-rag"
        ],
        "summary": "Update Notification Preferences",
        "operationId": "update_notification_preferences_api_v1_notifications_preferences_put",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationPreferenceUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPreferenceView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/process-due": {
      "post": {
        "tags": [
          "safebill-rag"
        ],
        "summary": "Process Due Notifications",
        "operationId": "process_due_notifications_api_v1_notifications_process_due_post",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationProcessResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/analytics": {
      "get": {
        "tags": [
          "safebill-rag"
        ],
        "summary": "Notifications Analytics",
        "operationId": "notifications_analytics_api_v1_notifications_analytics_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 30,
              "title": "Days"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationAnalyticsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NotificationAnalyticsResponse": {
        "properties": {
          "windowDays": {
            "type": "integer",
            "title": "Windowdays"
          },
          "totalAttempts": {
            "type": "integer",
            "title": "Totalattempts"
          },
          "successfulDeliveries": {
            "type": "integer",
            "title": "Successfuldeliveries"
          },
          "failedDeliveries": {
            "type": "integer",
            "title": "Faileddeliveries"
          },
          "deadLettered": {
            "type": "integer",
            "title": "Deadlettered"
          },
          "successRate": {
            "type": "number",
            "title": "Successrate"
          },
          "openRate": {
            "type": "number",
            "title": "Openrate"
          },
          "clickRate": {
            "type": "number",
            "title": "Clickrate"
          }
        },
        "type": "object",
        "required": [
          "windowDays",
          "totalAttempts",
          "successfulDeliveries",
          "failedDeliveries",
          "deadLettered",
          "successRate",
          "openRate",
          "clickRate"
        ],
        "title": "NotificationAnalyticsResponse"
      },
      "NotificationItem": {
        "properties": {
          "notificationId": {
            "type": "string",
            "title": "Notificationid"
          },
          "docId": {
            "type": "string",
            "title": "Docid"
          },
          "userId": {
            "type": "string",
            "title": "Userid"
          },
          "channel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel"
          },
          "eventType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eventtype"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "triggerAt": {
            "type": "string",
            "title": "Triggerat"
          },
          "readAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Readat"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "notificationId",
          "docId",
          "userId",
          "type",
          "title",
          "message",
          "triggerAt",
          "status"
        ],
        "title": "NotificationItem"
      },
      "NotificationPreferenceUpdateRequest": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 320
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "locale": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Locale"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          },
          "in_app_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "In App Enabled"
          },
          "email_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email Enabled"
          },
          "sms_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sms Enabled"
          },
          "sms_number": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Sms Number"
          },
          "push_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Push Enabled"
          },
          "whatsapp_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Whatsapp Enabled"
          },
          "whatsapp_number": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Whatsapp Number"
          },
          "alert_days": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alert Days"
          },
          "claim_alert_days": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claim Alert Days"
          }
        },
        "type": "object",
        "title": "NotificationPreferenceUpdateRequest"
      },
      "NotificationPreferenceView": {
        "properties": {
          "userId": {
            "type": "string",
            "title": "Userid"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "fullName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fullname"
          },
          "locale": {
            "type": "string",
            "title": "Locale",
            "default": "en"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone",
            "default": "UTC"
          },
          "inAppEnabled": {
            "type": "boolean",
            "title": "Inappenabled",
            "default": true
          },
          "emailEnabled": {
            "type": "boolean",
            "title": "Emailenabled",
            "default": true
          },
          "smsEnabled": {
            "type": "boolean",
            "title": "Smsenabled",
            "default": false
          },
          "smsNumber": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Smsnumber"
          },
          "pushEnabled": {
            "type": "boolean",
            "title": "Pushenabled",
            "default": true
          },
          "whatsappEnabled": {
            "type": "boolean",
            "title": "Whatsappenabled",
            "default": false
          },
          "whatsappNumber": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Whatsappnumber"
          },
          "alertDays": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Alertdays"
          },
          "claimAlertDays": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Claimalertdays"
          },
          "updatedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "userId",
          "email"
        ],
        "title": "NotificationPreferenceView"
      },
      "NotificationProcessResult": {
        "properties": {
          "processed": {
            "type": "integer",
            "title": "Processed"
          },
          "sent": {
            "type": "integer",
            "title": "Sent"
          },
          "failed": {
            "type": "integer",
            "title": "Failed"
          },
          "deadLettered": {
            "type": "integer",
            "title": "Deadlettered",
            "default": 0
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "processed",
          "sent",
          "failed"
        ],
        "title": "NotificationProcessResult"
      },
      "NotificationsResponse": {
        "properties": {
          "notifications": {
            "items": {
              "$ref": "#/components/schemas/NotificationItem"
            },
            "type": "array",
            "title": "Notifications"
          }
        },
        "type": "object",
        "required": [
          "notifications"
        ],
        "title": "NotificationsResponse"
      },
...
      "ReminderView": {
        "properties": {
          "reminderId": {
            "type": "string",
            "title": "Reminderid"
          },
          "docId": {
            "type": "string",
            "title": "Docid"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "triggerAt": {
            "type": "string",
            "title": "Triggerat"
          },
          "triggerType": {
            "type": "string",
            "title": "Triggertype"
          },
          "deliveryChannels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Deliverychannels"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "reminderId",
          "docId",
          "title",
          "triggerAt",
          "triggerType",
          "deliveryChannels",
          "status"
        ],
        "title": "ReminderView"
      },
      "RemindersResponse": {
        "properties": {
          "reminders": {
            "items": {
              "$ref": "#/components/schemas/ReminderView"
            },
            "type": "array",
            "title": "Reminders"
          }
        },
        "type": "object",
        "required": [
          "reminders"
        ],
        "title": "RemindersResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}